+++ module/application/src/main/kotlin/com/ecommerce/Application.kt
@@ -1,0 +1,11 @@
+package com.ecommerce
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class Application
+
+fun main(args: Array<String>) {
+    runApplication<Application>(*args)
+}
