--- module/application/src/main/kotlin/core/configuration/Security.kt
@@ -1,15 +1,0 @@
-package core.configuration
-
-import core.CommonPasswordEncoder
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-class Security {
-
-    @Bean
-    fun passwordEncoder(): PasswordEncoder {
-        return CommonPasswordEncoder()
-    }
-
-}
