--- module/core/exception/src/main/kotlin/BaseError.kt
@@ -1,4 +1,0 @@
-interface BaseError {
-    val code: String
-    val message: String
-}
