+++ module/core/exception/src/main/kotlin/com/ecommerce/BaseError.kt
@@ -1,0 +1,6 @@
+package com.ecommerce
+
+interface BaseError {
+    val code: String
+    val message: String
+}
