--- module/core/security/src/main/kotlin/PasswordEncoder.kt
@@ -1,4 +1,0 @@
-interface PasswordEncoder {
-    fun encode(raw: String): String
-    fun matches(raw: String, encoded: String): Boolean
-}
