+++ module/core/security/src/main/kotlin/com/ecommerce/PasswordEncoder.kt
@@ -1,0 +1,6 @@
+package com.ecommerce
+
+interface PasswordEncoder {
+    fun encode(raw: String): String
+    fun matches(raw: String, encoded: String): Boolean
+}
