+++ module/domain/address/src/main/kotlin/com/ecommerce/infrastructure/AddressJpaRepository.kt
@@ -1,0 +1,11 @@
+package com.ecommerce.infrastructure
+
+import org.springframework.data.jpa.repository.JpaRepository
+import service.Address
+import java.util.UUID
+
+interface AddressJpaRepository : JpaRepository<Address, UUID> {
+
+    fun findAllByMemberId(memberId: UUID): MutableList<Address>
+
+}
