+++ module/domain/address/src/main/kotlin/com/ecommerce/service/AddressCommand.kt
@@ -1,0 +1,10 @@
+package com.ecommerce.service
+
+class AddressCommand {
+    data class Create(
+        val country: String,
+        val city: String,
+        val street: String,
+        val detail: String,
+    )
+}
