--- module/domain/address/src/main/kotlin/infrastructure/AddressJpaRepository.kt
@@ -1,11 +1,0 @@
-package infrastructure
-
-import org.springframework.data.jpa.repository.JpaRepository
-import service.Address
-import java.util.UUID
-
-interface AddressJpaRepository : JpaRepository<Address, UUID> {
-
-    fun findAllByMemberId(memberId: UUID): MutableList<Address>
-
-}
