--- module/domain/address/src/main/kotlin/service/AddressCommand.kt
@@ -1,10 +1,0 @@
-package service
-
-class AddressCommand {
-    data class Create(
-        val country: String,
-        val city: String,
-        val street: String,
-        val detail: String,
-    )
-}
