+++ module/domain/member/src/main/kotlin/com/ecommerce/service/MemberEvent.kt
@@ -1,0 +1,10 @@
+package com.ecommerce.service
+
+import java.util.UUID
+
+class MemberEvent {
+
+    data class ChargeCash(val uuid: UUID, val amount: Long)
+    data class UseCash(val uuid: UUID, val amount: Long)
+
+}
