+++ module/domain/member/src/main/kotlin/com/ecommerce/service/MemberRepository.kt
@@ -1,0 +1,12 @@
+package com.ecommerce.service
+
+import java.util.UUID
+
+interface MemberRepository {
+
+    fun find(uuid: UUID): Member?
+    fun find(email: String, encryptedPassword: String): Member?
+    fun save(member: Member): Member
+    fun delete(member: Member)
+
+}
