--- module/domain/member/src/main/kotlin/service/MemberEvent.kt
@@ -1,10 +1,0 @@
-package service
-
-import java.util.UUID
-
-class MemberEvent {
-
-    data class ChargeCash(val uuid: UUID, val amount: Long)
-    data class UseCash(val uuid: UUID, val amount: Long)
-
-}
